Provides functionality needed for seats (#2120)
* Add ability to get an invite

* Provides error propagation while adding member to organization

+++ wsmaster/codenvy-hosted-api-organization/src/main/java/com/codenvy/organization/spi/MemberDao.java
@@ -119,5 +119,5 @@ public interface MemberDao {
      * @throws ServerException
      *         when any other error occurs during organizations fetching
      */
-    Page<OrganizationImpl> getOrganizations(String userId, int maxItems, int skipCount) throws ServerException;
+    Page<OrganizationImpl> getOrganizations(String userId, int maxItems, long skipCount) throws ServerException;
 }
